+++ src/extension.ts
@@ -1,18 +1,18 @@
 import * as vscode from "vscode";
-import { ConversionOptions, ConvertCallback, MatcherCallback } from "./types";
-import { forEachLineIn, forEachWordIn, insertLineComment } from "./lib";
+import { ConvertCallback, MatcherCallback } from "./types";
+import { forEachLineIn, forEachWordIn } from "./lib";
 import {
   hexStringToASCIIString,
   hexStringToBinaryString,
   hexStringToDecimalString,
 } from "./converters";
 import { isHexString } from "./matchers";
+import { insertLineComment } from "./comments";
 
-export function traverseWords(
+export function comment(
   editor: vscode.TextEditor | undefined,
   convert: ConvertCallback,
   matcher: MatcherCallback,
-  options: ConversionOptions,
 ) {
   if (!editor) {
     // No open editor
@@ -26,24 +26,71 @@ export function traverseWords(
 
   return editor.edit((builder) => {
     for (const selection of editor.selections) {
-      forEachLineIn(editor, selection, (line, lineIndex) => {
+      forEachLineIn(editor, selection, (line) => {
+        // Keeps track of mutable text and end index
+        let lineText: string = line.text;
+        let lineEndIndex: number = line.range.end.character;
+
+        forEachWordIn(editor, selection, line, (word) => {
+          if (matcher(word)) {
+            const converted = convert(word);
+            if (typeof converted === "string") {
+              insertLineComment(
+                line,
+                lineText,
+                lineEndIndex,
+                converted,
+                (comment, updatedLineText, updatedLineEndIndex) => {
+                  // Update line text and end index for next iteration.
+                  // Needed for multiple values on same line.
+                  lineText = updatedLineText;
+                  lineEndIndex = updatedLineEndIndex;
+
+                  builder.insert(
+                    new vscode.Position(line.lineNumber, lineEndIndex),
+                    comment,
+                  );
+                },
+              );
+            }
+          }
+        });
+      });
+    }
+  });
+}
+
+export function replace(
+  editor: vscode.TextEditor | undefined,
+  convert: ConvertCallback,
+  matcher: MatcherCallback,
+) {
+  if (!editor) {
+    // No open editor
+    return;
+  }
+
+  if (editor.selections.length < 1) {
+    // No selections
+    return;
+  }
+
+  return editor.edit((builder) => {
+    for (const selection of editor.selections) {
+      forEachLineIn(editor, selection, (line) => {
         forEachWordIn(editor, selection, line, (word, wordIndex) => {
           if (matcher(word)) {
             const converted = convert(word);
             if (typeof converted === "string") {
-              if (options.comment) {
-                insertLineComment(builder, line, converted);
-              } else {
-                builder.replace(
-                  new vscode.Range(
-                    lineIndex,
-                    wordIndex,
-                    lineIndex,
-                    wordIndex + word.length,
-                  ),
-                  converted,
-                );
-              }
+              builder.replace(
+                new vscode.Range(
+                  line.lineNumber,
+                  wordIndex,
+                  line.lineNumber,
+                  wordIndex + word.length,
+                ),
+                converted,
+              );
             }
           }
         });
@@ -57,42 +104,30 @@ export function activate(context: vscode.ExtensionContext) {
 
   context.subscriptions.push(
     vscode.commands.registerCommand("hex-multi-converter.hexToDecimal", () => {
-      traverseWords(editor, hexStringToDecimalString, isHexString, {
-        comment: false,
-      });
+      replace(editor, hexStringToDecimalString, isHexString);
     }),
     vscode.commands.registerCommand(
       "hex-multi-converter.commentHexAsDecimal",
       () => {
-        traverseWords(editor, hexStringToDecimalString, isHexString, {
-          comment: true,
-        });
+        comment(editor, hexStringToDecimalString, isHexString);
       },
     ),
     vscode.commands.registerCommand("hex-multi-converter.hexToASCII", () => {
-      traverseWords(editor, hexStringToASCIIString, isHexString, {
-        comment: false,
-      });
+      replace(editor, hexStringToASCIIString, isHexString);
     }),
     vscode.commands.registerCommand(
       "hex-multi-converter.commentHexAsASCII",
       () => {
-        traverseWords(editor, hexStringToASCIIString, isHexString, {
-          comment: true,
-        });
+        comment(editor, hexStringToASCIIString, isHexString);
       },
     ),
     vscode.commands.registerCommand("hex-multi-converter.hexToBinary", () => {
-      traverseWords(editor, hexStringToBinaryString, isHexString, {
-        comment: false,
-      });
+      replace(editor, hexStringToBinaryString, isHexString);
     }),
     vscode.commands.registerCommand(
       "hex-multi-converter.commentHexAsBinary",
       () => {
-        traverseWords(editor, hexStringToBinaryString, isHexString, {
-          comment: true,
-        });
+        comment(editor, hexStringToBinaryString, isHexString);
       },
     ),
   );