+++ src/test/suite/extension.test.ts
@@ -1,7 +1,7 @@
 import * as assert from "assert";
 import * as vscode from "vscode";
 
-import { traverseWords } from "../../extension";
+import { replace, comment } from "../../extension";
 import { hexStringToDecimalString } from "../../converters";
 import { isHexString } from "../../matchers";
 
@@ -21,9 +21,7 @@ suite("extension", () => {
 
     editor.selections = [new vscode.Selection(0, 0, 0, 10)];
 
-    await traverseWords(editor, hexStringToDecimalString, isHexString, {
-      comment: false,
-    });
+    await replace(editor, hexStringToDecimalString, isHexString);
 
     assert.strictEqual(editor.document.getText(), "Hello 255");
   });
@@ -33,9 +31,7 @@ suite("extension", () => {
 
     editor.selections = [new vscode.Selection(0, 5, 0, 9)];
 
-    await traverseWords(editor, hexStringToDecimalString, isHexString, {
-      comment: false,
-    });
+    await replace(editor, hexStringToDecimalString, isHexString);
 
     assert.strictEqual(editor.document.getText(), "0x00 255");
   });
@@ -45,9 +41,7 @@ suite("extension", () => {
 
     editor.selections = [new vscode.Selection(0, 0, 0, 10)];
 
-    await traverseWords(editor, hexStringToDecimalString, isHexString, {
-      comment: true,
-    });
+    await comment(editor, hexStringToDecimalString, isHexString);
 
     assert.strictEqual(editor.document.getText(), "Hello 0xff // 255");
   });
@@ -57,10 +51,21 @@ suite("extension", () => {
 
     editor.selections = [new vscode.Selection(0, 5, 0, 9)];
 
-    await traverseWords(editor, hexStringToDecimalString, isHexString, {
-      comment: true,
-    });
+    await comment(editor, hexStringToDecimalString, isHexString);
 
     assert.strictEqual(editor.document.getText(), "0x00 0xff // 255");
+  });
+
+  test("hexToDecimal - create comment for multiple values selected on one line", async () => {
+    const editor = await createEditor("0x00 0xff 0x10 0x20 0x1d foo");
+
+    editor.selections = [new vscode.Selection(0, 5, 0, 19)];
+
+    await comment(editor, hexStringToDecimalString, isHexString);
+
+    assert.strictEqual(
+      editor.document.getText(),
+      "0x00 0xff 0x10 0x20 0x1d foo // 255 16 32",
+    );
   });
 });