+++ src/types.ts
@@ -1,9 +1,9 @@
-import * as vscode from "vscode";
-
 export type ConvertCallback = (word: string) => string | null;
 
 export type MatcherCallback = (word: string) => boolean;
 
-export interface ConversionOptions {
-  comment: boolean;
-}
+export type InsertLineCommentCallback = (
+  comment: string,
+  updatedLineText: string,
+  updatedLineEndIndex: number,
+) => void;